requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.5
django-nine==0.2.5
djangorestframework==3.13.1
elastic-transport==8.4.0
elasticsearch==7.17.6
elasticsearch-dsl==7.4.0
idna==3.3
pydocstyle==6.1.1
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.2
urllib3==1.26.12
